change package name & remove db package
